fix: table width
